At the fun level: removed eq and ne and added isequal for equality tests.
At the chebfun level:
Better implementation of eq, so things like 
	x = chebfun(@(x) x);
	plot(max(x,0) == x)
now work.
Made the definition of ne.m consistent with the definition of eq.m.
New method not.m: Try
	f = ~(x==0); 
	f([-1:0.1:1])
New method isequal.m for equality tests for chebfuns.
Modified some of the tests because of the new definition is eq.m