First pass at Guide 10, on nonlinear IVPs and ODEs.  I am eager for Asgeir to look at this now.  Hopefully the next pass in a day or so will be ready for Toby.  Then on Tuesday I'll show it to the group at our regular meeting.
Executive summary: nonlinear backslash is working fine, but
since it's a pure Newton iteration, it's fragile.  We need to
robustify it.   I presume the key things to improve are the
step direction and the step length; I hope Asgeir and I can
have a chat with Nick Gould on this this week.