Deleted solve and solve_newton_damped, both types of Newton iterations (pure and damped) are now done in the new function solve_bvp_routines.
Various changes done to add robustness, findguess is now better to spot whether we just need to add another column to the quasimatrix to get an legit initial guess or whether something else is going wrong (e.g. division by zero).   

The mldivide function of linop now can now take three arguments, which saves us from having to change the eps in chebfunpref in the nonlinear backslash. Changing the eps in chebfunpref would cause problems if user aborted the run of programs, since the lower tolerance would remain in chebfunpref but not resetted to the old value.