Bug fix in chebfun/restrict for assigning imps.
Fix in scaling chebpts, legpts, and jacpts to infinite intervals.
(This is not allowed in jacpts).

Fix for turning warnings on and off in vectorcheck. (It was not resetting properly).

Increased default number of points in mesh and surf to 201, and added an option for the user to change this.