Modification to the GUI so we can now work with coupled systems.
This also means that we've enabled IVPs and BVPs of higher order than 2 (at least in theory, I haven't had much time to try it out).

Changes in relevant files for the GUI, as well as a typo fix in nonlinearBVPs.m.

Added a two new demos for the GUI, one is a coupled system and the other is an IVP.