Further merging of BVP and PDE GUIS into one GUI, chebde.
Added the functionality that when user toggles between modes, a new example is loaded (i.e. BVP example when user selects BVP mode and vice versa for PDEs).

Added function --folderchange-- into chebde.m, which makes it cleaner to switch between folder in the gui (this will probably only be of interest to Nick H).

Changed the demo button to a popup menu, and moved the loading of that popup menu into a seperate function (loaddemos.m).

Grouped the mode switching buttons into a panel.

Added a stop button for the BVP solving (tells the solver to stop after the current iteration and return the most recent solution).