Some late night, last minute fixes which should get the svn counter one step closer to 1111.
Fixed a bug in solve_bvp_routines which caused failure in when nonlinear backslash was called with all numerical nonzero RHS for systems, i.e. N\[0 2] (but N\[0 x] did work).

Added documentation for the chebgui which is shown when the help button is pressed.

Improved error handling. The parse function now deals with errors in more systematic way. If the parser throws an error (which means that the user is not using accepted syntax), or if the linop backslash throws one (which is most likely to happen if it fails to converge), an error dialog is brought up informing the user about the error. Other errors get rethrown since the error stack there gives more informative messages than just showing the error message in a dialogue.

Restored the version check in chebtest back such that versions 7.4 or later allow chebtest to run.