Fix in feval to prevent the following:
>> x = chebfun('x');
>> f = x.*(x-1);
>> f(1+1i)
ans =
     0

The problem was that chebfun feval looked for 
strict inequalities when choosing which fun
to evluate on. If the real part of the evaluation
point was on a breakpoint, it would fall through
the gaps. Additionally it would not be treated
as an impulse, as this tests for equality.

I fixed it by looking for ends(i) <= x < ends(i+1)
and evaluating at fun(i).