New examples structure with a few examples.
make_example is a routine for automatically publishing 
the examples, as well as producing html outputs for 
the user, and optionally .shtml outputs for the Chebfun
website.

The example files should be placed in one of the existing
subdirectories of $chebfunroot/examples/, and should begin
with the line
%% BRIEF DESCRIPTION OF EXAMPLE
(note the double %% and capitalisation).