Introducing the new chebgui class, which gets rid of the guifiles folder and creates a new @chebgui folder instead.
This is still fairly experimental, but should allow people to start experimenting with it (as well as enabling Nick H to implement ideas he's been working on for the GUI).

The GUI itself lies in the files chebguiwindow.m and chebguiwindow.fig. The name for these files are temporary, any suggestions welcome.