Added support for exporting the final time solution of a PDE from GUI to workspace in 'u_final'.
Exporting to .mat file uses the variable names as given in the GUI (before it 
defaulted to 'u'). Exporting to workspace also now uses the variable names in
the GUI as the default (rather than 'u').

Renamed "timedomain" to "time domain" when exporting PDEs.

Fixes in how the equations are ordered in PDE mode. (The kth equation must
be the one with the time derivative on the kth variable in @(u,v,w,...)
 - i.e., alphabetical).