Fully moved to new OOP, including storage of coefficients.
As expected, there's a bit of a hit in terms of performace,
of about 25%, which you can see by comparing the chebtest 
times here:
   http://www2.maths.ox.ac.uk/chebfun/files/OOP_times.html
or here if you want to play with the numbers
   http://www2.maths.ox.ac.uk/chebfun/files/OOP_times.ods   
I expect we should be able to get some of this back with
a few tweaks herre and there. (You'll notice one test 
that suffers particularly is scribbleio, so that's a good
place to start looking...)

To reiterate, MATLAB 7.4 is no longer supported.