Adding a new 'chebconst' class, which is a subclass of chebfun.
In order to automatically differeniate scalars, we need to store
jacobian information, which is imposible in doubles. To get around
this, when we wish to AD a function, say s = sum(u), which would
usually return a scalar, we instead set a flag on u which tells 
s to return a 'chebconst', rather than a double. Since chebconst
is a subclass of chebfun, it maintains the corredct jacobian info.