Partial fix to #216.
Define was calling vertcat once for each pair
of vertical concatenations, which is slow. 
Now they are stored in a cell and concatenated
all at once at the end. This kind of approach 
may be a way to overcome the problem of superslow
vertical fun concatenation in the new OOP.

>> tic, abs(chebfun('sin(80*pi*x)')); toc
Elapsed time is 3.348057 seconds.