Moving @domain class to the new OOP.
This removes the ability to construct linear chebfuns at the same 
time as a domain. For example, 
>> [d x] = domain(-1,1);
would need to be replaced with
>> d = domain(-1,1);
>> x = chebfun('x',d);
This is because the new OOP doesn't support multiple outputs from
class constructors. However, this is no great loss.

Updated chebtests and a couple of chebfun methods to reflect this.
Note that many of the Examples and the Guide will also need to be
updated.

Also, where possible, calls to domain have been replaced with 
vectors, as outlined in #112.