Adding CHEBFUN/INTEGRAL method to follow new MATLAB syntax.
Imorived a number of minor problems in the GUI lexer to
enables this with optionally 1 or 3 input arguments
(this was also broken for SUM, FEVAL, FRED, and VOLT).

Added a chebtest to test that the LEXER and PARSEr are
working correctly.

Fixed a minor bug in tidyInputs for piecewise functions.

All chebtests pass.