Added NH and AT's new asymptotic method of LEGPTS, which can be demonstrated to be around 30x faster than 'GLR' (at least in Matlab) and more accurate - see forthcoming paper for details. It is now the default method when N >= 100.
The method uses an asymptotic formula for the evaluation
of Legendre polynomials in a Newton iteration to find
the roots and derivatives values of P_N. The expansion 
is not valid near +/-1, so for now we are reverting to 
GLR in a VERY small region near the end points. We're 
exploring alternatives however.

We hope to add a corresponding JACPTS soon, as well
as LEGPOLY and JACPOLY using related ideas. We're
not sure yet whether LAGPTS and HERMPTS can also be
done in the same way.