Fix bug in feval() for evaluating chebfuns with many funs at matrix arguments.
Calling feval() for a chebfun with more than 10 funs causes a call to MATLAB's
intersect() function with the evaluation point passed as an argument.  If the
evaluation point is a matrix, intersect() fails.  Since the matrix structure of
the evaluation point is unimportant for its use in this call, regarding it as a
vector with the x(:) syntax can be used to fix the problem.