Attempted fix at catastrophic transpose error in r2008a. (Which somehoe manages to cause segfaults...)
The problem appears to be a bug in Matlab's BUILTIN() function,
which fails to call the built-in method. In transpose this is
necessary for switching the order of the indicies in a row
Chebfun.

The fix works by ensureing that all quasimatrices are indexed
by a single index, which works because Matlab is loose about 
indexing columsn or rows in this same way. 

This would not work for quasi-tensors, but fortunately they
don't exist. However, we should bear this in mind for cheb-
matrices.

I'll have to wait until the mornign to see if this works..