Fix bug in roots() for complex-valued chebfuns.
Mr. Hrothgar observed that roots() finds a root for the following everywhere
nonzero complex-valued chebfun:

  >> f_r = chebfun(@(x) x, [0 1]);
  >> f_i = chebfun(@(x) 0.5*sin(1000*x) + 1.5, [0 1]);
  >> f = f_r + 1i*f_i;
  >> r = roots(f)
  r =
       0

Of course, f is not actually zero there:

  >> f(r)
  ans =
         0.000000000000000 + 1.500000000000000i

The reason for this is that we take real parts of coefficients in a couple of
places within the rootsunit_coeffs() subroutine.  Removing the calls to real()
fixes the problem.